Ransomeware/requirements.txt
pycryptodome==3.9.2
mysql-connector-python
pyinstaller
configparser==4.0.2